+++ examples/resources/elasticstack_elasticsearch_data_stream/import.sh
@@ -1,0 +1,2 @@
+terraform import elasticstack_elasticsearch_data_stream.my_data_stream <cluster_uuid>/<data_stream_name>
+